+++ internal/provider/data_source_iosxe_bgp_address_family_ipv4.go
@@ -143,6 +143,42 @@ func (d *BGPAddressFamilyIPv4DataSource) Schema(ctx context.Context, req datasou
 					},
 				},
 			},
+			"admin_distances": schema.ListNestedAttribute{
+				MarkdownDescription: "",
+				Computed:            true,
+				NestedObject: schema.NestedAttributeObject{
+					Attributes: map[string]schema.Attribute{
+						"distance": schema.Int64Attribute{
+							MarkdownDescription: "",
+							Computed:            true,
+						},
+						"source_ip": schema.StringAttribute{
+							MarkdownDescription: "",
+							Computed:            true,
+						},
+						"wildcard": schema.StringAttribute{
+							MarkdownDescription: "",
+							Computed:            true,
+						},
+						"acl": schema.StringAttribute{
+							MarkdownDescription: "",
+							Computed:            true,
+						},
+					},
+				},
+			},
+			"distance_bgp_external": schema.Int64Attribute{
+				MarkdownDescription: "",
+				Computed:            true,
+			},
+			"distance_bgp_internal": schema.Int64Attribute{
+				MarkdownDescription: "",
+				Computed:            true,
+			},
+			"distance_bgp_local": schema.Int64Attribute{
+				MarkdownDescription: "",
+				Computed:            true,
+			},
 		},
 	}
 }